plop

+++ Lead-It/Assets/Script/Country.cs
@@ -14,7 +14,7 @@ public class Country : MonoBehaviour
     private string		CountryName;
     private decimal		Budget;
     private float		DayRatio;
-	private decimal		Popularity;
+	private float		Popularity;
 	private decimal		DayPast;
 	// INVISIBLE
     private bool		IsDead;
